+++ src/global/components/Dashboard/BasicInfo/DashboardAvatar/DashboardAvatar.scss
