+++ src/global/components/Dashboard/DashInterviewQ/DeletePost.scss
