+++ src/global/components/Dashboard/DeleteModal/DeleteModal.js
