+++ src/global/components/Dashboard/DeleteModal/DeleteModal.scss
