+++ src/global/components/Dashboard/DeleteModal/index.js
